+++ ft_memchr.c
@@ -6,7 +6,7 @@
 /*   By: kmuhlbau <kmuhlbau@student.42heilbronn.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/17 14:26:25 by kmuhlbau          #+#    #+#             */
-/*   Updated: 2024/10/07 23:46:13 by kmuhlbau         ###   ########.fr       */
+/*   Updated: 2024/10/11 14:02:07 by kmuhlbau         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 void	*ft_memchr(const void *s, int c, size_t n)
 {
 	const unsigned char	*ptr;
-	unsigned int		i;
+	size_t				i;
 
 	i = 0;
 	ptr = s;
