+++ ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: kmuhlbau <kmuhlbau@student.42heilbronn.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/15 20:11:03 by kmuhlbau          #+#    #+#             */
-/*   Updated: 2024/10/08 11:08:21 by kmuhlbau         ###   ########.fr       */
+/*   Updated: 2024/10/11 14:09:04 by kmuhlbau         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_strncmp(const char *s1, const char *s2, size_t len)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	if (len == 0)
