+++ ft_strnstr.c
@@ -6,7 +6,7 @@
 /*   By: kmuhlbau <kmuhlbau@student.42heilbronn.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/15 19:07:40 by kmuhlbau          #+#    #+#             */
-/*   Updated: 2024/10/07 23:12:06 by kmuhlbau         ###   ########.fr       */
+/*   Updated: 2024/10/11 14:09:47 by kmuhlbau         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,12 +14,12 @@
 
 char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
 {
-	unsigned int	i;
-	unsigned int	j;
+	size_t	i;
+	size_t	j;
 
 	i = 0;
 	if (ft_strlen(needle) == 0)
-		return ((char *) haystack);
+		return ((char *)haystack);
 	while (haystack[i] != '\0' && i < len)
 	{
 		if (haystack[i] == needle[0])
@@ -29,7 +29,7 @@ char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
 					+ j) < len)
 				j++;
 			if (needle[j] == '\0')
-				return ((char *) &haystack[i]);
+				return ((char *)&haystack[i]);
 		}
 		i++;
 	}
